requirements.txt
markovify==0.9.4
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
python-dotenv==0.20.0
requests==2.28.0
tweepy==4.10.0